requirements.txt
aiohttp==3.7.4
altgraph==0.17
APScheduler==3.7.0
astroid==2.5.6
async-timeout==3.0.1
attrs==20.3.0
certifi==2022.12.7
chardet==3.0.4
colorama==0.4.4
discord.py==2.0.1
dnspython==2.1.0
future==0.18.3
idna==3.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
numpy==1.23.3
pandas==1.2.1
pefile==2022.5.30
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pylint==2.8.3
pymongo==4.3.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==6.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
tzlocal==2.1
wrapt==1.12.1
yarl==1.6.3
